change around platform specific components, more migration stuff
When Pulsar is working enough, i want to rewrite all of this. At least retype it. That way, we cannot miss anything and are _sure_ stuff is up to date. The current stuff are only meant to be an "archival" copy, and really for me to mess around with, setup vuepress, and get familliar with it