reorganise some things
shove as much things into docs as possible, keep it all contained in that folder? be consistent at least. before this, most things were in docs, but the vuepress config and vuepress client config were in the root which now that i'm thinking about it kindof didn't make much sense